Will continue to work in the weekend. Spend most of my energy on that Accesibility Resit assignment.

+++ TreeMarket_Klas4_Groep7/Models/Product.cs
@@ -20,7 +20,7 @@ namespace TreeMarket_Klas4_Groep7.Models
 
         //Een op meer relatie
         [ForeignKey(nameof(Leverancier))]
-        public decimal LeverancierID { get; set; }
+        public int LeverancierID { get; set; }
         public Leverancier Leverancier { get; set; }
     }
 }
